% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_compartment_optimization.R
\name{run_compartment_optimization}
\alias{run_compartment_optimization}
\title{Run optimization for individual samples}
\usage{
run_compartment_optimization(
  data,
  compartments_n,
  samples_n,
  genes_n,
  resultH,
  resultW,
  parallel_n = 2
)
}
\arguments{
\item{data}{Gene expression target data, a matrix-like object. The rows should represent genes, and each row must have a unique row name. Each column should represent a different sample.}

\item{compartments_n}{The factorization rank (number of factors) to be used during NMF. This function argument should be a positive integer value.}

\item{samples_n}{The number of samples in the dataset}

\item{genes_n}{The number of genes in the dataset}

\item{resultH}{The H matrix from (FaStaNMF) deconvolution}

\item{parallel_n}{The number of cores available for parallel optimization - will drastically impact runtime}
}
\value{
A list containing W matrix for every sample; in each matrix, rows are genes and columns are deconvolved compartments/factors/tissues types
}
\description{
\code{run_compartment_optimization} will perform gene-specific sample-specific optimization on the output of FastaNFM or other deconvolution technique
}
