% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSCD.R
\name{SSCD}
\alias{SSCD}
\title{Sample Specific Compartment Deconvolution}
\usage{
SSCD(
  data,
  compartments_n = 3,
  nrun = 200,
  nmf_seed = 124578,
  mvg = 1000,
  parallel_n = 2,
  ...
)
}
\arguments{
\item{data}{Gene expression target data, a matrix-like object. The rows should represent genes, and each row must have a unique row name. Each column should represent a different sample.}

\item{compartments_n}{The factorization rank (number of factors) to be used during NMF. This function argument should be a positive integer value.}

\item{nrun}{The desired number of FaStaNMF runs to be run}

\item{nmf_seed}{The desired seed to be used for NMF}

\item{mvg}{A numerical argument determining how many of the most variable genes to look at during the first steps of FaStaNMF.}

\item{...}{Other arguments to be passed to FaStaNMF.}
}
\value{
A list containing W matrix for every sample; in each matrix, rows are genes and columns are deconvolved compartments/factors/tissues types
}
\description{
\code{SSCD} will perform sample specific gene expression deconvolution on a matrix or data frame of gene expression data using FaStaNMF and gene-specific sample-specific optimization.
}
